move tests to own files
