+++ app_spec/test/files/entry.js
@@ -1,0 +1,368 @@
+module.exports = scenario => {
+
+    scenario('create_tagged_post and retrieve all tags', async (s, t, { alice }) => {
+        const result1 = await alice.app.callSync("blog", "create_tagged_post", {
+          content: "Tutorial on amazing Holochain design patterns",
+          tag: "work"
+        })
+        t.ok(result1.Ok)
+      
+        const result2 = await alice.app.callSync("blog", "create_tagged_post", {
+          content: "Fly tying, is it for you?",
+          tag: "fishing"
+        })
+        t.ok(result2.Ok)
+      
+        const getResult = await alice.app.callSync("blog", "my_posts", {})
+        t.equal(getResult.Ok.links.length, 2)
+        let tags = getResult.Ok.links.map(l => l.tag)
+        t.ok(tags.includes("work"))
+        t.ok(tags.includes("fishing"))
+      })
+      
+      scenario('create_tagged_post and retrieve exact tag match', async (s, t, { alice }) => {
+        const result1 = await alice.app.callSync("blog", "create_tagged_post", {
+          content: "Tutorial on amazing Holochain design patterns",
+          tag: "work"
+        })
+        t.ok(result1.Ok)
+      
+        const result2 = await alice.app.callSync("blog", "create_tagged_post", {
+          content: "Fly tying, is it for you?",
+          tag: "fishing"
+        })
+        t.ok(result2.Ok)
+      
+        const getResult = await alice.app.callSync("blog", "my_posts", {tag: "fishing"})
+        t.equal(getResult.Ok.links.length, 1)
+        let tags = getResult.Ok.links.map(l => l.tag)
+        t.notOk(tags.includes("work"))
+        t.ok(tags.includes("fishing"))
+      })
+      
+      scenario('create_tagged_post and retrieve regex tag match', async (s, t, { alice }) => {
+        const result1 = await alice.app.callSync("blog", "create_tagged_post", {
+          content: "A post made on the 10th of October",
+          tag: "10/10/2019"
+        })
+        t.ok(result1.Ok)
+      
+        const result2 = await alice.app.callSync("blog", "create_tagged_post", {
+          content: "A post made on the 10th of September",
+          tag: "10/9/2019"
+        })
+        t.ok(result2.Ok)
+      
+        const getResult = await alice.app.callSync("blog", "my_posts", {tag: "^10\/[0-9]+\/2019$"})
+        t.equal(getResult.Ok.links.length, 2)
+        let tags = getResult.Ok.links.map(l => l.tag)
+        t.ok(tags.includes("10/10/2019"))
+        t.ok(tags.includes("10/9/2019"))
+      })
+
+      scenario('delete_entry_post', async (s, t, { alice, bob }) => {
+        const content = "Hello Holo world 321"
+        const in_reply_to = null
+        const params = { content, in_reply_to }
+      
+        //commit create_post
+        const createResult = await alice.app.callSync("blog", "create_post", params)
+      
+        t.ok(createResult.Ok)
+      
+      
+        //delete entry post
+        const deletionParams = { post_address: createResult.Ok }
+        const deletionResult = await alice.app.callSync("blog", "delete_entry_post", deletionParams)
+      
+        t.ok(deletionResult.Ok)
+      
+      
+        //delete should fail
+        const failedDelete = await bob.app.callSync("blog", "delete_entry_post", { post_address: createResult.Ok })
+        t.deepEqual(failedDelete.Err, { Internal: 'Entry Could Not Be Found' });
+      
+        //get initial entry
+        const GetInitialParamsResult = await bob.app.call("blog", "get_initial_post", { post_address: createResult.Ok })
+        t.deepEqual(JSON.parse(GetInitialParamsResult.Ok.App[1]), { content: "Hello Holo world 321", date_created: "now" });
+      
+        const entryWithOptionsGet = { post_address: createResult.Ok }
+        const entryWithOptionsGetResult = await bob.app.call("blog", "get_post_with_options", entryWithOptionsGet);
+        t.deepEqual(JSON.parse(entryWithOptionsGetResult.Ok.result.All.items[0].entry.App[1]), { content: "Hello Holo world 321", date_created: "now" })
+      })
+      
+      scenario('update_entry_validation', async (s, t, { alice, bob }) => {
+        //update entry does not exist
+        const updateParams = { post_address: "1234", new_content: "Hello Holo V2" }
+        const UpdateResult = await bob.app.callSync("blog", "update_post", updateParams)
+      
+        t.deepEqual(UpdateResult.Err, { Internal: 'failed to update post' });
+      
+        const content = "Hello Holo world 321"
+        const in_reply_to = null
+        const params = { content, in_reply_to }
+      
+        //commit create_post
+        const createResult = await alice.app.callSync("blog", "create_post", params)
+      
+        t.ok(createResult.Ok)
+      
+        const updateParamsV2 = { post_address: createResult.Ok, new_content: "Hello Holo world 321" }
+        const UpdateResultV2 = await bob.app.callSync("blog", "update_post", updateParamsV2)
+        t.deepEqual(JSON.parse(UpdateResultV2.Err.Internal).kind.ValidationFailed, "Trying to modify with same data");
+      
+      
+      })
+      
+      scenario('update_post', async (s, t, { alice, bob }) => {
+        const content = "Hello Holo world 123"
+        const in_reply_to = null
+        const params = { content, in_reply_to }
+      
+        //commit version 1
+        const createResult = await alice.app.callSync("blog", "create_post", params)
+        t.ok(createResult.Ok)
+      
+        //get v1
+        const updatedPostV1 = await alice.app.call("blog", "get_post", { post_address: createResult.Ok })
+        const UpdatePostV1Content = { content: "Hello Holo world 123", date_created: "now" };
+        t.ok(updatedPostV1.Ok)
+        t.deepEqual(JSON.parse(updatedPostV1.Ok.App[1]), UpdatePostV1Content)
+      
+        //update to version 2
+        const updatePostContentV2 = { content: "Hello Holo V2", date_created: "now" };
+        const updateParamsV2 = { post_address: createResult.Ok, new_content: "Hello Holo V2" }
+        const UpdateResultV2 = await bob.app.callSync("blog", "update_post", updateParamsV2)
+        t.ok(UpdateResultV2.Ok)
+        t.notOk(UpdateResultV2.Err)
+      
+        //get v2 using initial adderss
+        const updatedPostv2Initial = await alice.app.call("blog", "get_post", { post_address: createResult.Ok })
+        t.ok(updatedPostv2Initial.Ok)
+        t.notOk(updatedPostv2Initial.Err)
+        t.deepEqual(JSON.parse(updatedPostv2Initial.Ok.App[1]), updatePostContentV2)
+      
+        //get v2 latest address
+        const updatedPostv2Latest = await alice.app.call("blog", "get_post", { post_address: UpdateResultV2.Ok })
+        t.ok(updatedPostv2Latest.Ok)
+        t.notOk(updatedPostv2Latest.Err)
+        t.deepEqual(JSON.parse(updatedPostv2Latest.Ok.App[1]), updatePostContentV2)
+      
+      
+        //get v1 using initial address
+        const GetInitialPostV1Initial = await alice.app.call("blog", "get_initial_post", { post_address: createResult.Ok })
+        t.ok(GetInitialPostV1Initial.Ok)
+        t.notOk(GetInitialPostV1Initial.Err)
+        t.deepEqual(JSON.parse(GetInitialPostV1Initial.Ok.App[1]), { content: "Hello Holo world 123", date_created: "now" })
+      
+        //get v2 latest address
+        const GetInitialPostV2Latest = await alice.app.call("blog", "get_initial_post", { post_address: UpdateResultV2.Ok })
+        t.ok(GetInitialPostV2Latest.Ok)
+        t.notOk(GetInitialPostV2Latest.Err)
+        t.deepEqual(JSON.parse(GetInitialPostV2Latest.Ok.App[1]), updatePostContentV2)
+      
+        //update to version 3
+        const UpdatePostV3Content = { content: "Hello Holo V3", date_created: "now" };
+        const updateParamsV3 = { post_address: createResult.Ok, new_content: "Hello Holo V3" }
+        const UpdateResultV3 = await alice.app.callSync("blog", "update_post", updateParamsV3)
+        t.ok(UpdateResultV3.Ok)
+        t.notOk(UpdateResultV3.Err)
+      
+        //get v3 using initial adderss
+        const updatedPostV3Initial = await alice.app.call("blog", "get_post", { post_address: createResult.Ok })
+        t.ok(updatedPostV3Initial.Ok)
+        t.notOk(updatedPostV3Initial.Err)
+        t.deepEqual(JSON.parse(updatedPostV3Initial.Ok.App[1]), UpdatePostV3Content)
+      
+        //get v3 using address of v2
+        const updatedPostV3Latest = await alice.app.call("blog", "get_post", { post_address: UpdateResultV2.Ok })
+        t.ok(updatedPostV3Latest.Ok)
+        t.notOk(updatedPostV3Latest.Err)
+        t.deepEqual(JSON.parse(updatedPostV3Latest.Ok.App[1]), UpdatePostV3Content)
+      
+        //update to version 4
+        const updatePostV4Content = { content: "Hello Holo V4", date_created: "now" };
+        const updateParamsV4 = { post_address: createResult.Ok, new_content: "Hello Holo V4" }
+        const UpdateResultV4 = await alice.app.callSync("blog", "update_post", updateParamsV4)
+        t.notOk(UpdateResultV4.Err)
+        t.ok(UpdateResultV4.Ok)
+      
+        //get history entry v4
+        const entryHistoryV4Params = { post_address: UpdateResultV4.Ok }
+        const entryHistoryV4 = await alice.app.call("blog", "get_history_post", entryHistoryV4Params)
+        t.ok(UpdateResultV4.Ok)
+        t.notOk(UpdateResultV4.Err)
+        t.deepEqual(entryHistoryV4.Ok.items.length, 1);
+        t.deepEqual(JSON.parse(entryHistoryV4.Ok.items[0].entry.App[1]), updatePostV4Content);
+        t.deepEqual(entryHistoryV4.Ok.items[0].meta.address, UpdateResultV4.Ok);
+        t.deepEqual(entryHistoryV4.Ok.items[0].meta.crud_status, "live");
+      
+        //get history entry all
+        const entryHistoryAllParams = { post_address: createResult.Ok }
+        const entryHistoryAll = await alice.app.call("blog", "get_history_post", entryHistoryAllParams)
+      
+        t.deepEqual(entryHistoryAll.Ok.items.length, 4);
+        t.deepEqual(JSON.parse(entryHistoryAll.Ok.items[0].entry.App[1]), { content: "Hello Holo world 123", date_created: "now" });
+        t.deepEqual(entryHistoryAll.Ok.items[0].meta.address, createResult.Ok);
+        t.deepEqual(entryHistoryAll.Ok.items[0].meta.crud_status, "modified");
+        t.deepEqual(entryHistoryAll.Ok.crud_links[createResult.Ok], UpdateResultV2.Ok)
+      
+        t.deepEqual(JSON.parse(entryHistoryAll.Ok.items[1].entry.App[1]), updatePostContentV2);
+        t.deepEqual(entryHistoryAll.Ok.items[1].meta.address, UpdateResultV2.Ok);
+        t.deepEqual(entryHistoryAll.Ok.items[1].meta.crud_status, "modified");
+        t.deepEqual(entryHistoryAll.Ok.crud_links[UpdateResultV2.Ok], UpdateResultV3.Ok)
+      
+        t.deepEqual(JSON.parse(entryHistoryAll.Ok.items[2].entry.App[1]), UpdatePostV3Content);
+        t.deepEqual(entryHistoryAll.Ok.items[2].meta.address, UpdateResultV3.Ok);
+        t.deepEqual(entryHistoryAll.Ok.items[2].meta.crud_status, "modified");
+        t.deepEqual(entryHistoryAll.Ok.crud_links[UpdateResultV3.Ok], UpdateResultV4.Ok)
+      
+        t.deepEqual(JSON.parse(entryHistoryAll.Ok.items[3].entry.App[1]), updatePostV4Content);
+        t.deepEqual(entryHistoryAll.Ok.items[3].meta.address, UpdateResultV4.Ok);
+        t.deepEqual(entryHistoryAll.Ok.items[3].meta.crud_status, "live");
+        t.notOk(entryHistoryAll.Ok.crud_links[UpdateResultV4.Ok])
+      
+        const entryWithOptionsGet = { post_address: createResult.Ok }
+        const entryWithOptionsGetResult = await alice.app.call("blog", "get_post_with_options_latest", entryWithOptionsGet);
+      
+        t.deepEqual(JSON.parse(entryWithOptionsGetResult.Ok.App[1]), updatePostV4Content);
+      })
+
+      scenario('remove_update_modifed_entry', async (s, t, { alice, bob }) => {
+        const content = "Hello Holo world 123"
+        const in_reply_to = null
+        const params = { content, in_reply_to }
+      
+        //commit version 1
+        const createResult = await alice.app.callSync("blog", "create_post", params)
+        t.ok(createResult.Ok)
+        //get entry
+        const updatedPostV1 = await alice.app.call("blog", "get_post", { post_address: createResult.Ok })
+        t.ok(updatedPostV1.Ok)
+        t.deepEqual(JSON.parse(updatedPostV1.Ok.App[1]), { content: "Hello Holo world 123", date_created: "now" })
+      
+        //delete
+        const removeParamsV2 = { post_address: createResult.Ok }
+        const removeResultV2 = await bob.app.callSync("blog", "delete_entry_post", removeParamsV2)
+        t.ok(removeResultV2.Ok)
+      
+        //get v2 using initial adders
+        const Postv2Initial = await alice.app.call("blog", "get_initial_post", { post_address: createResult.Ok })
+        t.ok(Postv2Initial.Ok)
+        t.deepEqual(JSON.parse(Postv2Initial.Ok.App[1]), { content: "Hello Holo world 123", date_created: "now" })
+      
+        //failed delete
+        const failedDelete = await alice.app.callSync("blog", "delete_entry_post", { post_address: createResult.Ok })
+        t.deepEqual(failedDelete.Err, { Internal: 'Entry Could Not Be Found' });
+      })
+
+
+      scenario('get sources', async (s, t, { alice, bob, carol }) => {
+        const params = { content: 'whatever', in_reply_to: null }
+        const address = await alice.app.callSync('blog', 'create_post', params).then(x => x.Ok)
+        const address1 = await alice.app.callSync('blog', 'create_post', params).then(x => x.Ok)
+        const address2 = await bob.app.callSync('blog', 'create_post', params).then(x => x.Ok)
+        const address3 = await carol.app.callSync('blog', 'create_post', params).then(x => x.Ok)
+        t.equal(address, address1)
+        t.equal(address, address2)
+        t.equal(address, address3)
+        const sources1 = (await alice.app.call('blog', 'get_sources', { address })).Ok.sort()
+        const sources2 = (await bob.app.call('blog', 'get_sources', { address })).Ok.sort()
+        const sources3 = (await carol.app.call('blog', 'get_sources', { address })).Ok.sort()
+        // NB: alice shows up twice because she published the same entry twice
+        const expected = [alice.app.agentId, alice.app.agentId, bob.app.agentId, carol.app.agentId].sort()
+        t.deepEqual(sources1, expected)
+        t.deepEqual(sources2, expected)
+        t.deepEqual(sources3, expected)
+      })
+
+      scenario('scenario test create & publish post -> get from other instance', async (s, t, { alice, bob }) => {
+
+        const initialContent = "Holo world"
+        const params = { content: initialContent, in_reply_to: null }
+        const create_result = await alice.app.callSync("blog", "create_post", params)
+      
+        const params2 = { content: "post 2", in_reply_to: null }
+        const create_result2 = await bob.app.callSync("blog", "create_post", params2)
+      
+        t.equal(create_result.Ok.length, 46)
+        t.equal(create_result.Ok, "QmY6MfiuhHnQ1kg7RwNZJNUQhwDxTFL45AAPnpJMNPEoxk")
+      
+        const post_address = create_result.Ok
+        const params_get = { post_address }
+      
+        const result = await bob.app.call("blog", "get_post", params_get)
+        const value = JSON.parse(result.Ok.App[1])
+        t.equal(value.content, initialContent)
+      })
+      
+
+      scenario('get_post with non-existant address returns null', async (s, t, { alice }) => {
+
+        const post_address = "RANDOM"
+        const params_get = { post_address }
+        const result = await alice.app.call("blog", "get_post", params_get)
+      
+        // should be Ok value but null
+        // lookup did not error
+        // successfully discovered the entry does not exity
+        const entry = result.Ok
+        t.same(entry, null)
+      })
+
+      scenario('create/get_post roundtrip', async (s, t, { alice }) => {
+        const content = "Holo world"
+        const in_reply_to = null
+        const params = { content, in_reply_to }
+        const create_post_result = await alice.app.call("blog", "create_post", params)
+        const post_address = create_post_result.Ok
+      
+        const params_get = { post_address }
+        const result = await alice.app.call("blog", "get_post", params_get)
+      
+        const entry_value = JSON.parse(result.Ok.App[1])
+        t.comment("get_post() entry_value = " + entry_value + "")
+        t.equal(entry_value.content, content)
+        t.equal(entry_value.date_created, "now")
+      
+      })
+
+      scenario('hash_post', async (s, t, { alice }) => {
+
+        const params = { content: "Holo world" }
+        const result = await alice.app.call("blog", "post_address", params)
+      
+        t.equal(result.Ok, "QmY6MfiuhHnQ1kg7RwNZJNUQhwDxTFL45AAPnpJMNPEoxk")
+      })
+
+      scenario('create_post', async (s, t, { alice }) => {
+
+        const content = "Holo world"
+        const in_reply_to = null
+        const params = { content, in_reply_to }
+        const result = await alice.app.call("blog", "create_post", params)
+      
+        t.ok(result.Ok)
+        t.notOk(result.Err)
+        t.equal(result.Ok, "QmY6MfiuhHnQ1kg7RwNZJNUQhwDxTFL45AAPnpJMNPEoxk")
+      })
+
+      scenario('post max content size 280 characters', async (s, t, { alice }) => {
+
+        const content = "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum."
+        const in_reply_to = null
+        const params = { content, in_reply_to }
+        const result = await alice.app.call("blog", "create_post", params)
+      
+        // result should be an error
+        t.ok(result.Err);
+        t.notOk(result.Ok)
+      
+        const inner = JSON.parse(result.Err.Internal)
+      
+        t.ok(inner.file)
+        t.deepEqual(inner.kind, { "ValidationFailed": "Content too long" })
+        t.ok(inner.line)
+      })
+
+    }
